VolatileOwningFloatUnsafe extends VolatileFloatAccess
